Port fastSign to the server-wallet
fastSign depends on libauth, and libauth makes building with webpack
complicated, leading to broken build steps on circle.

server-wallet will never use webpack, so this solves this problem.
