+++ packages/nitro-protocol/package.json
@@ -6,7 +6,6 @@
   "browser": "dist/nitro-protocol.min.js",
   "bugs": "https://github.com/statechannels/monorepo/issues",
   "dependencies": {
-    "@bitauth/libauth": "1.17.1",
     "@openzeppelin/contracts": "3.0.1",
     "ethers": "4.0.45"
   },
