+++ packages/nitro-protocol/src/signatures.ts
@@ -1,12 +1,8 @@
 import {Wallet, utils} from 'ethers';
-import {instantiateSecp256k1, Secp256k1} from '@bitauth/libauth';
 
 import {hashChallengeMessage} from './contract/challenge';
 import {getChannelId} from './contract/channel';
 import {hashState, State} from './contract/state';
-
-let secp256k1: Secp256k1;
-export const initialized: Promise<any> = instantiateSecp256k1().then(m => (secp256k1 = m));
 
 // This is the same as the ethers Signature type
 // But we redefine it here to prevent the below issue
@@ -52,21 +48,6 @@ export function signState(state: State, privateKey: string): SignedState {
   return {state, signature};
 }
 
-export async function fastSignState(
-  state: State,
-  privateKey: string
-): Promise<{state: State; signature: string}> {
-  const wallet = new Wallet(privateKey);
-  if (state.channel.participants.indexOf(wallet.address) < 0) {
-    throw new Error("The state must be signed with a participant's private key");
-  }
-
-  const hashedState = hashState(state);
-
-  const signature = await fastSignData(hashedState, privateKey);
-  return {state, signature};
-}
-
 export async function sign(wallet: Wallet, msgHash: string | Uint8Array) {
   // MsgHash is a hex string
   // Returns an object with v, r, and s properties.
@@ -103,22 +84,9 @@ export function signChallengeMessage(
 function hashMessage(hashedData: string): string {
   return utils.hashMessage(utils.arrayify(hashedData));
 }
+
 function signData(hashedData: string, privateKey: string): utils.Signature {
   const signingKey = new utils.SigningKey(privateKey);
 
   return utils.splitSignature(signingKey.signDigest(hashMessage(hashedData)));
-}
-
-export async function fastSignData(hashedData: string, privateKey: string): Promise<string> {
-  await initialized;
-
-  const hash = hashMessage(hashedData);
-  const digest = Buffer.from(hash.substr(2), 'hex');
-  const signature = secp256k1.signMessageHashRecoverableCompact(
-    Buffer.from(privateKey.substr(2), 'hex'),
-    digest
-  );
-
-  const v = 27 + signature.recoveryId;
-  return '0x' + Buffer.from(signature.signature).toString('hex') + v.toString(16);
 }
