+++ packages/nitro-protocol/test/src/signatures.test.ts
@@ -4,32 +4,10 @@ import {arrayify, splitSignature, verifyMessage} from 'ethers/utils';
 
 import {hashChallengeMessage} from '../../src/contract/challenge';
 import {hashState, State} from '../../src/contract/state';
-import {
-  getStateSignerAddress,
-  signChallengeMessage,
-  signState,
-  fastSignState,
-} from '../../src/signatures';
+import {getStateSignerAddress, signChallengeMessage, signState} from '../../src/signatures';
 
 describe('signatures', () => {
   describe('signState', () => {
-    it('sign vs fastSign', async () => {
-      const wallet = Wallet.createRandom();
-      const state: State = {
-        channel: {chainId: '0x1', channelNonce: 0x01, participants: [wallet.address]},
-        outcome: [],
-        turnNum: 1,
-        isFinal: false,
-        appData: '0x0',
-        appDefinition: AddressZero,
-        challengeDuration: 0x5,
-      };
-
-      const signedState = signState(state, wallet.privateKey);
-      const fastSignedState = fastSignState(state, wallet.privateKey);
-      expect(utils.joinSignature(signedState.signature)).toEqual((await fastSignedState).signature);
-    });
-
     it('signs a state', async () => {
       const wallet = Wallet.createRandom();
       const state: State = {
