+++ packages/server-wallet/package.json
@@ -7,6 +7,7 @@
     "@statechannels/client-api-schema": "^0.3.0",
     "@statechannels/nitro-protocol": "^0.3.0",
     "@statechannels/wallet-core": "^0.3.0",
+    "@bitauth/libauth": "1.17.1",
     "ethers": "5.0.7",
     "fp-ts": "2.7.0",
     "knex": "0.20.9",
