+++ packages/server-wallet/tsconfig.json
@@ -2,6 +2,8 @@
   "extends": "../../tsconfig.json",
   "compilerOptions": {
     "allowSyntheticDefaultImports": true,
+    // ["es2017", "esnext.bigint", "dom"] needed for https://github.com/bitauth/libauth#typescript-types
+    "lib": ["es5", "es6", "ES2017", "esnext.bigint", "dom"],
     "target": "es2019",
     "module": "commonjs",
     "outDir": "lib",
